Add composition to improve the SceneDelegate
